second commit

+++ src/main/java/it/solvingteam/olympics/repository/AthleteRepository.java
@@ -4,6 +4,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import it.solvingteam.olympics.model.athlete.Athlete;
 
-public interface AthleteRepository extends JpaRepository<Athlete, Integer>{
+public interface AthleteRepository extends JpaRepository<Athlete, Long>{
 
 }
